requirements.txt
Flask==3.0.0
pytest==7.4.3
pyyaml==6.0.1
requests==2.31.0
jsonschema==4.20.0
